Update README.md
This is a list of all Super Bowl championship matches in the last 61 years. After I added all the statistics and analytics in the ReadMe file, I was able to see that that was duplicated on the desktop app as well which will eventually get added to the public repository.